2010-08-23  chl  <chl@tuxfamily.org>
	* src/CVS/Root:
	* regress/CVS/Root:
	- update

	* openbsd-compat/sys-tree.h:
	* src/file.h:
	- add "openbsd-compat/sys-tree.h" header

	* src/diff3.c (diff3_getline):
	- rename getline() to diff3_getline(), to avoid colision with
	POSIX getline().

	* src/client.c:
	* src/diff_internals.c:
	* src/import.c:
	* src/log.c:
	* src/server.c:
	* src/xmalloc.c:
	- add #define _GNU_SOURCE to properly define vasprintf()
	  prototype

	* src/remote.c:
	- cast to follow format string

+++ src/remote.c
@@ -184,7 +184,7 @@ cvs_remote_send_file(const char *path, int _fd)
 	cvs_modetostr(st.st_mode, buf, sizeof(buf));
 	cvs_remote_output(buf);
 
-	(void)xsnprintf(buf, sizeof(buf), "%lld", st.st_size);
+	(void)xsnprintf(buf, sizeof(buf), "%lld", (long long int) st.st_size);
 	cvs_remote_output(buf);
 
 	if ((in = fdopen(fd, "r")) == NULL)
@@ -204,7 +204,9 @@ cvs_remote_send_file(const char *path, int _fd)
 	}
 
 	if (total != st.st_size)
-		fatal("length mismatch, %lld vs %lld", total, st.st_size);
+		fatal("length mismatch, %lld vs %lld",
+		      (long long int) total,
+		      (long long int) st.st_size);
 
 	(void)fclose(in);
 }
